requirements.txt
google-genai
pinecone
python-dotenv
openai
pypdf
flask
plotly
scikit-learn
numpy
